Features - Implement Auto-Suggestion - SQL Editor (#49)
* feat: Add monaco-sql-languages package

This commit adds the "monaco-sql-languages" package to the project's dependencies. This package provides support for SQL language features in the Monaco editor. It is added to the "ui/package.json" file.

* refactor: Update monaco-sql-languages package and add completion item provider

This commit refactors the code in the editor component to update the monaco-sql-languages package and add a completion item provider for SQL language. The completion item provider provides suggestions for auto-completion in the editor. The implementation currently includes suggestions for auto-suggestions, but the tables, columns, and tables with columns suggestions are yet to be implemented. This commit also removes the unused code related to tables and columns suggestions.

* refactor: :fire: remove unused editor hook

* refactor: :recycle: remove duplicate suggestions in autoSuggestionCompletionItems

* refactor: :recycle: update autocomplete suggestions in editor